requirements_boox.txt
anthropic>=0.39.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
